[CL-3174] Delete dead code
