--- front/app/components/admin/PostManager/components/LazyPostPreview/Idea/tracks.ts
@@ -1,8 +1,0 @@
-export default {
-  titleProfanityError: {
-    name: 'Profane idea title blocked',
-  },
-  descriptionProfanityError: {
-    name: 'Profane idea description blocked',
-  },
-};
